禁言某个房间的某个人

+++ models/archive.go
@@ -45,13 +45,13 @@ type Event struct {
 
 //RoomConf 聊天室配置
 type RoomConf struct {
-	MaxOnline       int              //最大在线人数
-	Silence         bool             //(除管理员)全员禁言
-	SpeakNotAllowed map[float64]bool //被禁言人员
+	MaxOnline       int          //最大在线人数
+	Silence         bool         //(除管理员)全员禁言
+	SpeakNotAllowed map[int]bool //被禁言人员
 }
 
 //User 用户信息
 type User struct {
-	UserID   float64
+	UserID   int
 	UserType string
 }
